Fixing code in notebook as per reported errors
Multiple Github issues logged against the tutorial which features this notebook reveal an issue in the ordering of these code blocks. Changing the install block to occur before the declaration of variables addresses the bug. Without doing this, the second to last code block doesn't recognize the "storage_account" variable. Also, the uninstall step causes a blocking error that prevents the rest of the notebook from running. Based on an issue that was logged to this repo, it looks like another contributor found a workaround of using "!" instead of "?". I included that workaround in this commit.